✨ feat(auth): improve signup form with firstName/lastName fields
- Merge SignUpForm component into page for simpler structure
- Add firstName and lastName as required fields (displayed side-by-side)
- Reorder form fields: firstName/lastName → username → email → password
- Configure additional fields in better-auth with proper type inference
- Map username to better-auth's name field (stored in database username column)
- Add type-safe plugin (inferAdditionalFields) for client-side field definitions
- Update i18n translations for new fields (zh-CN, en-US)
