+++ locales/en-US/auth.json
@@ -56,6 +56,8 @@
     "errors": {
       "emailInvalid": "Please enter a valid email address",
       "emailRequired": "Please enter your email address",
+      "firstNameRequired": "Please enter your first name",
+      "lastNameRequired": "Please enter your last name",
       "passwordMinLength": "Password must be at least 8 characters",
       "passwordRequired": "Please enter your password",
       "usernameRequired": "Please enter your username"
@@ -63,7 +65,9 @@
     "signup": {
       "emailPlaceholder": "Enter your email address",
       "error": "Sign up failed, please try again",
+      "firstNamePlaceholder": "First Name",
       "hasAccount": "Already have an account?",
+      "lastNamePlaceholder": "Last Name",
       "passwordPlaceholder": "Enter your password (at least 8 characters)",
       "signinLink": "Sign in now",
       "submit": "Sign Up",
