+++ src/libs/better-auth/auth-client.ts
@@ -1,5 +1,7 @@
+import { inferAdditionalFields } from 'better-auth/client/plugins';
 import { createAuthClient } from 'better-auth/react';
 
+import type { auth } from '@/auth';
 import { getAuthConfig } from '@/envs/auth';
 
 const { NEXT_PUBLIC_BETTER_AUTH_URL } = getAuthConfig();
@@ -7,4 +9,5 @@ const { NEXT_PUBLIC_BETTER_AUTH_URL } = getAuthConfig();
 export const { signIn, signUp, useSession } = createAuthClient({
   /** The base URL of the server (optional if you're using the same domain) */
   baseURL: NEXT_PUBLIC_BETTER_AUTH_URL,
+  plugins: [inferAdditionalFields<typeof auth>()],
 });
