+++ src/locales/default/auth.ts
@@ -56,6 +56,8 @@ export default {
     errors: {
       emailInvalid: '请输入有效的邮箱地址',
       emailRequired: '请输入邮箱地址',
+      firstNameRequired: '请输入名字',
+      lastNameRequired: '请输入姓氏',
       passwordMinLength: '密码至少需要 8 个字符',
       passwordRequired: '请输入密码',
       usernameRequired: '请输入用户名',
@@ -63,7 +65,9 @@ export default {
     signup: {
       emailPlaceholder: '请输入邮箱地址',
       error: '注册失败，请重试',
+      firstNamePlaceholder: '名字',
       hasAccount: '已有账号？',
+      lastNamePlaceholder: '姓氏',
       passwordPlaceholder: '请输入密码（至少 8 个字符）',
       signinLink: '立即登录',
       submit: '注册',
